追加: BMI一覧ページ作成 :#88

+++ db/migrate/20210625152713_create_bmis.rb
@@ -1,7 +1,7 @@
 class CreateBmis < ActiveRecord::Migration[5.2]
   def change
     create_table :bmis do |t|
-      t.integer :status
+      t.float :status
       t.date :record_on
       t.references :user, foreign_key: true
